+++ protocols/relay/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Remove duplicated forwarding of pending events to connection handler.
 - Emit `relay::Event::ReservationClosed` when an active reservation is dropped due to the connection closing.
   See [PR 5869](https://github.com/libp2p/rust-libp2p/pull/5869).
+- Include denial reason in `relay::Event::CircuitReqDenied` and `relay::Event::ReservationReqDenied`
+  See [PR 6067](https://github.com/libp2p/rust-libp2p/pull/6067/files).
 
 ## 0.19.0